refactor: add custom error handling

+++ src/inverted_index/constants.rs
@@ -1,3 +1,3 @@
-pub(super) const BOLD_WEIGHT: f64 = 3.0;
-pub(super) const HEADER_WEIGHT: f64 = 5.0;
-pub(super) const TITLE_WEIGHT: f64 = 10.0;
+pub(super) const BOLD_WEIGHT: f32 = 3.0;
+pub(super) const HEADER_WEIGHT: f32 = 5.0;
+pub(super) const TITLE_WEIGHT: f32 = 10.0;
